==> codes/00_ggplot_theme.R <==
#-----------------------------------------------------------------------------
# Set ggplot theme
#-----------------------------------------------------------------------------
library(ggplot2)
library(ggthemes)
library(ggthemes)
library(patchwork)
library(ggtext)
#-----------------------------------------------------------------------------
theme_set(
  #theme_clean() + 
  theme_classic() +
    theme(
      panel.background = element_rect(fill = "transparent"), # bg of the panel
      plot.background = element_rect(fill = "transparent", color = NA), # bg of the plot
      legend.background = element_rect(color = "white"),
      legend.box.background = element_rect(fill = "transparent"), # get rid of legend panel bg
      panel.grid.major = element_blank(), 
      panel.grid.minor = element_blank(),
      panel.spacing = unit(10, "lines"))
)
#---------------------------------------------------------------------------------------
